requirements.txt
python-dotenv~=1.0.1
web3~=7.5.0